server/requirements.txt
redis>=4.5.0 
celery>=5.3.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
django>=4.2.0